+++ features/transaction/buyerorder/src/main/java/com/tokopedia/buyerorder/detail/view/presenter/OrderListDetailPresenter.java
@@ -368,7 +368,7 @@ public class OrderListDetailPresenter extends BaseDaggerPresenter<OrderListDetai
         for (Items item : items) {
             JsonObject passenger = new JsonObject();
 
-            int productId = 0;
+            long productId = 0;
             int quantity = 0;
             int shopId = 0;
             String notes = "";
